move solve_cubic_equation into math subdirectory
