+++ src/clib/dust/multi_grain_species/calc_grain_size_increment_species_1d.cpp
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include "grackle.h"
-#include "solve_cubic_equation.hpp"
+#include "math/solve_cubic_equation.hpp"
 #include "utils-cpp.hpp"
 
 #include "calc_grain_size_increment_species_1d.hpp"
