+++ src/clib/math/solve_cubic_equation.cpp
