+++ tests/unit/CMakeLists.txt
@@ -21,13 +21,10 @@ add_executable(runLinAlgTests test_linalg.cpp)
 target_link_libraries(runLinAlgTests testdeps)
 gtest_discover_tests(runLinAlgTests)
 
-add_executable(runMathTests math/test_integrate.cpp)
+add_executable(runMathTests
+  math/test_integrate.cpp math/test_solve_cubic_equation.cpp)
 target_link_libraries(runMathTests testdeps)
 gtest_discover_tests(runMathTests)
-
-add_executable(runSolveCubicEquationTests test_solve_cubic_equation.cpp)
-target_link_libraries(runSolveCubicEquationTests testdeps)
-gtest_discover_tests(runSolveCubicEquationTests)
 
 add_executable(runGrainSpeciesIntoTests test_grain_species_info.cpp)
 target_link_libraries(runGrainSpeciesIntoTests testdeps)
